requirements.txt
alembic==1.7.5
certifi==2021.10.8
click==8.0.3
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
PyJWT==1.7.1
SQLAlchemy==1.4.27
Werkzeug==2.0.2
zipp==3.6.0
